test: update to run in serial and have login
