--- test/blog/delete-blog.test.js
@@ -1,53 +1,0 @@
-import tap from 'tap';
-import { build } from '../../src/app.js';
-import 'must/register.js';
-
-tap.mochaGlobals();
-
-const prefix = '/api';
-
-describe('Delete a blog should work', async () => {
-  let app;
-
-  before(async () => {
-    app = await build();
-  });
-
-  it('Should return the success = true if ID is deleted', async () => {
-    const newBlog = {
-      title: 'New Blog for get',
-      description: 'Some description'
-    };
-
-    const createResponse = await app.inject({
-      method: 'POST',
-      url: `${prefix}/blog`,
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(newBlog)
-    });
-
-    const { id } = await createResponse.json();
-
-    const response = await app.inject({
-      method: 'DELETE',
-      url: `${prefix}/blog/${id}`
-    });
-
-    // this checks if HTTP status code is equal to 200
-    response.statusCode.must.be.equal(200);
-
-    const result = await response.json();
-
-    // expect success is true
-    result.success.must.be.true();
-
-    const getResponse = await app.inject({
-      method: 'GET',
-      url: `${prefix}/blog/${id}`
-    });
-
-    getResponse.statusCode.must.be.equal(404);
-  });
-});
