+++ test/tap-parallel-not-ok/blog/update-blog.test.js
@@ -1,6 +1,9 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
+import Chance from 'chance';
+
+const chance = new Chance();
 
 tap.mochaGlobals();
 
@@ -11,6 +14,59 @@ describe('Update a blog should work', async () => {
 
   before(async () => {
     app = await build();
+  });
+
+  const newUser = {
+    username: chance.email({ domain: 'example.com' }),
+    password: chance.string({ length: 12 }),
+    firstName: chance.first(),
+    lastName: chance.last()
+  };
+
+  let cookie = '';
+
+  it('Should return the user that was created a new user', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/register`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(newUser)
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(200);
+
+    const result = await response.json();
+
+    // expect that id exists
+    result.username.must.be.equal(newUser.username);
+    result.firstName.must.be.equal(newUser.firstName);
+    result.lastName.must.be.equal(newUser.lastName);
+
+    // expect createdDate and updatedDate is not null
+    result.createdDate.must.not.be.null();
+    result.updatedDate.must.not.be.null();
+  });
+
+  it('Login should work', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        username: newUser.username,
+        password: newUser.password
+      })
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
   });
 
   it('Should update the object given an ID', async () => {
@@ -29,7 +85,8 @@ describe('Update a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -40,7 +97,8 @@ describe('Update a blog should work', async () => {
       method: 'PUT',
       url: `${prefix}/blog/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerBlog)
     });
@@ -76,7 +134,8 @@ describe('Update a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -87,7 +146,8 @@ describe('Update a blog should work', async () => {
       method: 'PUT',
       url: `${prefix}/blog/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerBlog)
     });
@@ -123,7 +183,8 @@ describe('Update a blog should work', async () => {
       method: 'POST',
       url: `${prefix}/blog`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newBlog)
     });
@@ -134,7 +195,8 @@ describe('Update a blog should work', async () => {
       method: 'PUT',
       url: `${prefix}/blog/${id}`,
       headers: {
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
+        cookie
       },
       body: JSON.stringify(newerBlog)
     });
