+++ test/tap-parallel-not-ok/user/login.test.js
@@ -1,0 +1,101 @@
+import tap from 'tap';
+import { build } from '../../../src/app.js';
+import 'must/register.js';
+import Chance from 'chance';
+
+const chance = new Chance();
+
+tap.mochaGlobals();
+
+const prefix = '/api';
+
+describe('Logging in a user should work', async () => {
+  let app;
+
+  before(async () => {
+    app = await build();
+  });
+
+  const newUser = {
+    username: chance.email({ domain: 'example.com' }),
+    password: chance.string({ length: 12 }),
+    firstName: chance.first(),
+    lastName: chance.last()
+  };
+
+  it('Should return the user that was created a new user', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/register`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify(newUser)
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(200);
+
+    const result = await response.json();
+
+    // expect that id exists
+    result.username.must.be.equal(newUser.username);
+    result.firstName.must.be.equal(newUser.firstName);
+    result.lastName.must.be.equal(newUser.lastName);
+
+    // expect createdDate and updatedDate is not null
+    result.createdDate.must.not.be.null();
+    result.updatedDate.must.not.be.null();
+  });
+
+  it('Login should work', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        username: newUser.username,
+        password: newUser.password
+      })
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(200);
+  });
+
+  it('Login should return an error if username doesn\'t exist', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        username: 'test',
+        password: 'password'
+      })
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(401);
+  });
+
+  it('Login should return an error if password is incorrect', async () => {
+    const response = await app.inject({
+      method: 'POST',
+      url: `${prefix}/login`,
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        username: newUser.username,
+        password: 'password'
+      })
+    });
+
+    // this checks if HTTP status code is equal to 200
+    response.statusCode.must.be.equal(401);
+  });
+});
