+++ test/tap-parallel-not-ok/user/logout.test.js
@@ -1,5 +1,5 @@
 import tap from 'tap';
-import { build } from '../../src/app.js';
+import { build } from '../../../src/app.js';
 import 'must/register.js';
 import Chance from 'chance';
 
@@ -9,7 +9,7 @@ tap.mochaGlobals();
 
 const prefix = '/api';
 
-describe('Register a user should work', async () => {
+describe('Logging out a user should work', async () => {
   let app;
 
   before(async () => {
@@ -22,6 +22,8 @@ describe('Register a user should work', async () => {
     firstName: chance.first(),
     lastName: chance.last()
   };
+
+  let cookie = '';
 
   it('Should return the user that was created a new user', async () => {
     const response = await app.inject({
@@ -48,17 +50,49 @@ describe('Register a user should work', async () => {
     result.updatedDate.must.not.be.null();
   });
 
-  it('Should return error HTTP code 400 if using the same username', async () => {
+  it('Login should work', async () => {
     const response = await app.inject({
       method: 'POST',
-      url: `${prefix}/register`,
+      url: `${prefix}/login`,
       headers: {
         'Content-Type': 'application/json'
       },
-      body: JSON.stringify(newUser)
+      body: JSON.stringify({
+        username: newUser.username,
+        password: newUser.password
+      })
     });
 
     // this checks if HTTP status code is equal to 200
-    response.statusCode.must.be.equal(400);
+    response.statusCode.must.be.equal(200);
+
+    cookie = response.headers['set-cookie'];
+  });
+
+  it('Logout should work', async () => {
+    const response = await app.inject({
+      method: 'GET',
+      url: `${prefix}/logout`,
+      headers: {
+        'Content-Type': 'application/json',
+        cookie
+      }
+    });
+
+    // this checks if HTTP status code is equal to 401
+    response.statusCode.must.be.equal(200);
+  });
+
+  it('Logout should return an error without a cookie', async () => {
+    const response = await app.inject({
+      method: 'GET',
+      url: `${prefix}/logout`,
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+
+    // this checks if HTTP status code is equal to 401
+    response.statusCode.must.be.equal(401);
   });
 });
